+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+Mox.defmock(X402.PaymentIdentifierCacheMock, for: X402.Extensions.PaymentIdentifier.Cache)
+
 ExUnit.start(exclude: [:smoke])